--- ui/server/graphql/resolvers/bucket/index.ts
@@ -1,1 +1,0 @@
-export * as queries from "./queries";
