+++ ui/server/graphql/resolvers/index.ts
@@ -1,4 +1,3 @@
-import slugify from "../../utils/slugify";
 //import liveUpdate from "../../services/liveUpdate.service";
 import prisma from "../../prisma";
 import { GraphQLScalarType } from "graphql";
@@ -7,25 +6,11 @@ import { GraphQLJSONObject } from "graphql-type-json";
 import { Kind } from "graphql/language";
 import dayjs from "dayjs";
 import { combineResolvers, skip } from "graphql-resolvers";
-import discourse from "../../lib/discourse";
-import {
-  allocateToMember,
-  bulkAllocate,
-  contribute,
-  getGroup,
-} from "../../controller";
-import subscribers from "../../subscribers/discourse.subscriber";
+import { contribute, getGroup } from "../../controller";
 import {
   bucketIncome,
   bucketMinGoal,
   bucketTotalContributions,
-  canViewRound,
-  getRoundMember,
-  getCurrentGroupAndMember,
-  getGroupMember,
-  isAndGetCollMember,
-  isAndGetCollMemberOrGroupAdmin,
-  isCollAdmin,
   isCollOrGroupAdmin,
   isGrantingOpen,
   roundMemberBalance,
@@ -34,59 +19,22 @@ import {
   bucketMaxGoal,
   getLanguageProgress,
 } from "./helpers";
-import emailService from "server/services/EmailService/email.service";
-import { RoundTransaction } from "server/types";
-import { sign, verify } from "server/utils/jwt";
-import { appLink } from "utils/internalLinks";
-import validateUsername from "utils/validateUsername";
-import { updateFundedPercentage } from "../resolvers/helpers";
-import {
-  createGroupInvitationLink,
-  deleteGroupInvitationLink,
-} from "./group/mutations";
 
 //queries
-import { queries as userQueries } from "./user";
-import { queries as groupQueries } from "./group";
-import { queries as roundQueries } from "./round";
-import { queries as bucketQueries } from "./bucket";
+import {
+  groupQueries,
+  userQueries,
+  roundQueries,
+  bucketQueries,
+} from "./queries";
 
-const { groupHasDiscourse, generateComment } = subscribers;
-
-const isCollMember = async (parent, { roundId, bucketId }, { user }) => {
-  if (!user) throw new Error("You need to be logged in");
-
-  const roundMember = await getRoundMember({
-    userId: user.id,
-    roundId,
-    bucketId,
-  });
-  // const roundMember = await prisma.roundMember.findUnique({
-  //   where: { userId_roundId: { userId: user.id, roundId } },
-  // });
-  if (!roundMember) {
-    throw new Error("Round member does not exist");
-  } else if (!roundMember.isApproved) {
-    throw new Error("Round member is not approved");
-  } else if (!roundMember.hasJoined) {
-    throw new Error("Round member has not accepted the invitation");
-  }
-
-  return skip;
-};
-
-const isCollModOrAdmin = async (parent, { bucketId, roundId }, { user }) => {
-  if (!user) throw new Error("You need to be logged in");
-  const roundMember = await getRoundMember({
-    userId: user.id,
-    bucketId,
-    roundId,
-  });
-
-  if (!(roundMember?.isModerator || roundMember?.isAdmin))
-    throw new Error("You need to be admin or moderator of the round");
-  return skip;
-};
+//mutations
+import {
+  userMutations,
+  groupMutations,
+  roundMutations,
+  bucketMutations,
+} from "./mutations";
 
 export const isGroupAdmin = async (parent, { groupId }, { user }) => {
   if (!user) throw new Error("You need to be logged in");
@@ -144,340 +92,11 @@ const resolvers = {
     },
   },
   Mutation: {
-    createGroup: async (parent, { name, slug, logo }, { user, eventHub }) => {
-      if (!user) throw new Error("You need to be logged in!");
+    ...userMutations,
+    ...groupMutations,
+    ...roundMutations,
+    ...bucketMutations,
 
-      const group = await prisma.group.create({
-        data: {
-          name,
-          slug: slugify(slug),
-          logo,
-          groupMembers: { create: { userId: user.id, isAdmin: true } },
-        },
-        include: {
-          groupMembers: true,
-        },
-      });
-
-      await eventHub.publish("create-group", {
-        currentGroup: group,
-        currentGroupMember: group.groupMembers[0],
-      });
-
-      return group;
-    },
-    editGroup: combineResolvers(
-      isGroupAdmin,
-      async (
-        parent,
-        { groupId, name, info, slug, registrationPolicy, logo, visibility },
-        { user, eventHub }
-      ) => {
-        if (name?.length === 0) throw new Error("Group name cannot be blank");
-        if (slug?.length === 0) throw new Error("Group slug cannot be blank");
-        if (info?.length > 500) throw new Error("Group info too long");
-        const group = await prisma.group.update({
-          where: {
-            id: groupId,
-          },
-          data: {
-            name,
-            info,
-            logo,
-            registrationPolicy,
-            visibility,
-            slug: slug !== undefined ? slugify(slug) : undefined,
-          },
-        });
-
-        // TODO: add back
-        // await eventHub.publish("edit-group", {
-        //   currentGroup: group,
-        //   currentGroupMember,
-        // });
-        return group;
-      }
-    ),
-    setTodosFinished: combineResolvers(
-      isGroupAdmin,
-      async (parent, { groupId }) => {
-        const group = await prisma.group.update({
-          where: { id: groupId },
-          data: { finishedTodos: true },
-        });
-        return group;
-      }
-    ),
-    createRound: async (
-      parent,
-      { groupId, slug, title, currency, registrationPolicy },
-      { user }
-    ) => {
-      let singleRound = false;
-      if (!groupId) {
-        let rootGroup = await prisma.group.findUnique({
-          where: { slug: "c" },
-        });
-        if (!rootGroup) {
-          rootGroup = await prisma.group.create({
-            data: { slug: "c", name: "Root" },
-          });
-        }
-        groupId = rootGroup.id;
-        singleRound = true;
-      } else {
-        await isGroupAdmin(null, { groupId }, { user });
-      }
-      const round = await prisma.round.create({
-        data: {
-          slug,
-          title,
-          currency,
-          registrationPolicy,
-          group: { connect: { id: groupId } },
-          singleRound,
-          statusAccount: { create: {} },
-          roundMember: {
-            create: {
-              user: { connect: { id: user.id } },
-              isAdmin: true,
-              isApproved: true,
-              statusAccount: { create: {} },
-              incomingAccount: { create: {} },
-              outgoingAccount: { create: {} },
-            },
-          },
-          fields: {
-            create: {
-              name: "Description",
-              description: "Describe your bucket",
-              type: "MULTILINE_TEXT",
-              isRequired: false,
-              position: 1001,
-            },
-          },
-        },
-      });
-
-      // await eventHub.publish("create-round", {
-      //   currentGroup,
-      //   currentGroupMember,
-      //   round: round,
-      // });
-
-      return round;
-    },
-    editRound: combineResolvers(
-      isCollOrGroupAdmin,
-      async (
-        parent,
-        {
-          roundId,
-          slug,
-          title,
-          archived,
-          registrationPolicy,
-          visibility,
-          info,
-          color,
-          about,
-          bucketReviewIsOpen,
-          discourseCategoryId,
-        }
-      ) => {
-        return prisma.round.update({
-          where: { id: roundId },
-          data: {
-            ...(slug && { slug: slugify(slug) }),
-            title,
-            archived,
-            registrationPolicy,
-            visibility,
-            info,
-            about,
-            color,
-            bucketReviewIsOpen,
-            discourseCategoryId,
-          },
-        });
-      }
-    ),
-
-    createGroupInvitationLink,
-    deleteGroupInvitationLink,
-
-    createInvitationLink: async (parent, { roundId }, { user }) => {
-      const isAdmin =
-        (await !!user) &&
-        isCollAdmin({
-          userId: user?.id,
-          roundId,
-        });
-
-      if (!isAdmin) {
-        throw new Error("You need to be admin to create invitation link");
-      }
-
-      const inviteNonce = Date.now();
-      const round = await prisma.round.update({
-        where: { id: roundId },
-        data: { inviteNonce },
-      });
-      return {
-        link: round.inviteNonce,
-      };
-    },
-    deleteInvitationLink: async (parent, { roundId }, { user }) => {
-      const isAdmin =
-        (await !!user) &&
-        isCollAdmin({
-          userId: user?.id,
-          roundId,
-        });
-
-      if (!isAdmin) {
-        throw new Error("You need to be admin to create delete link");
-      }
-
-      await prisma.round.update({
-        where: { id: roundId },
-        data: { inviteNonce: null },
-      });
-      return {
-        link: null,
-      };
-    },
-    joinInvitationLink: async (parent, { token }, { user }) => {
-      if (!user) {
-        throw new Error("You need to be logged in to join the group");
-      }
-
-      const payload = verify(token);
-
-      if (!payload) {
-        throw new Error("Invalid invitation link");
-      }
-
-      const { roundId, groupId, nonce: inviteNonce } = payload;
-
-      if (roundId) {
-        const round = await prisma.round.findFirst({
-          where: { id: roundId, inviteNonce },
-        });
-
-        if (!round) {
-          throw new Error("Round link expired");
-        }
-
-        const isApproved = true;
-        const roundMember = await prisma.roundMember.upsert({
-          where: { userId_roundId: { userId: user.id, roundId } },
-          create: {
-            round: { connect: { id: roundId } },
-            user: { connect: { id: user.id } },
-            isApproved,
-            statusAccount: { create: {} },
-            incomingAccount: { create: {} },
-            outgoingAccount: { create: {} },
-          },
-          update: { isApproved, hasJoined: true, isRemoved: false },
-        });
-
-        return { id: roundMember.id, roundId: roundMember.roundId };
-      } else {
-        const group = await prisma.group.findFirst({
-          where: { id: groupId, inviteNonce },
-        });
-
-        if (!group) {
-          throw new Error("Group invitation link expired");
-        }
-
-        const groupMember = await prisma.groupMember.create({
-          data: { userId: user.id, groupId: groupId },
-        });
-
-        return { id: groupMember.id, groupId: groupMember.groupId };
-      }
-    },
-    deleteRound: combineResolvers(
-      isCollOrGroupAdmin,
-      async (parent, { roundId }) =>
-        prisma.round.update({
-          where: { id: roundId },
-          data: { deleted: true },
-        })
-    ),
-    addGuideline: combineResolvers(
-      isCollOrGroupAdmin,
-      async (parent, { roundId, guideline: { title, description } }) => {
-        const guidelines = await prisma.guideline.findMany({
-          where: { roundId: roundId },
-        });
-
-        const position =
-          guidelines
-            .map((g) => g.position)
-            .reduce((a, b) => Math.max(a, b), 1000) + 1;
-
-        const guideline = await prisma.guideline.create({
-          data: { roundId: roundId, title, description, position },
-          include: { round: true },
-        });
-        return guideline.round;
-      }
-    ),
-    editGuideline: combineResolvers(
-      isCollOrGroupAdmin,
-      async (
-        parent,
-        { roundId, guidelineId, guideline: { title, description } }
-      ) => {
-        const round = await prisma.round.findUnique({
-          where: { id: roundId },
-          include: { guidelines: true },
-        });
-
-        if (!round.guidelines.map((g) => g.id).includes(guidelineId))
-          throw new Error("This guideline is not part of this round");
-
-        const guideline = await prisma.guideline.update({
-          where: { id: guidelineId },
-          data: { title, description },
-          include: { round: true },
-        });
-
-        return guideline.round;
-      }
-    ),
-    setGuidelinePosition: combineResolvers(
-      isCollOrGroupAdmin,
-      async (parent, { roundId, guidelineId, newPosition }, { user }) => {
-        const round = await prisma.round.findUnique({
-          where: { id: roundId },
-          include: { guidelines: true },
-        });
-
-        if (!round.guidelines.map((g) => g.id).includes(guidelineId))
-          throw new Error("This guideline is not part of this round");
-
-        const guideline = await prisma.guideline.update({
-          where: { id: guidelineId },
-          data: { position: newPosition },
-          include: { round: true },
-        });
-
-        return guideline.round;
-      }
-    ),
-    deleteGuideline: combineResolvers(
-      isCollOrGroupAdmin,
-      async (parent, { roundId, guidelineId }) =>
-        prisma.round.update({
-          where: { id: roundId },
-          data: { guidelines: { delete: { id: guidelineId } } },
-        })
-    ),
     addCustomField: combineResolvers(
       isCollOrGroupAdmin,
       async (
@@ -562,202 +181,7 @@ const resolvers = {
           data: { fields: { delete: { id: fieldId } } },
         })
     ),
-    createBucket: combineResolvers(
-      isCollMember,
-      async (parent, { roundId, title }, { user, eventHub }) => {
-        const round = await prisma.round.findUnique({
-          where: { id: roundId },
-          include: {
-            group: {
-              include: {
-                groupMembers: { where: { userId: user.id } },
-                discourse: true,
-              },
-            },
-          },
-        });
 
-        const currentGroupMember = round?.group?.groupMembers?.[0];
-
-        const bucketCreationIsOpen = round.bucketCreationCloses
-          ? dayjs().isBefore(dayjs(round.bucketCreationCloses))
-          : true;
-
-        if (!bucketCreationIsOpen)
-          throw new Error("Bucket creation is not open");
-
-        const bucket = await prisma.bucket.create({
-          data: {
-            round: { connect: { id: roundId } },
-            title,
-            statusAccount: { create: {} },
-            outgoingAccount: { create: {} },
-            cocreators: {
-              connect: {
-                userId_roundId: { userId: user.id, roundId },
-              },
-            },
-          },
-        });
-
-        await eventHub.publish("create-bucket", {
-          currentGroup: round.group,
-          currentGroupMember,
-          bucket: bucket,
-          round: round,
-        });
-
-        return bucket;
-      }
-    ),
-    editBucket: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (
-        parent,
-        {
-          bucketId,
-          title,
-          description,
-          summary,
-          images,
-          budgetItems,
-          directFundingEnabled,
-          directFundingType,
-          exchangeDescription,
-          exchangeMinimumContribution,
-          exchangeVat,
-        },
-        { user, eventHub }
-      ) => {
-        if (
-          exchangeMinimumContribution !== undefined &&
-          exchangeMinimumContribution < 0
-        ) {
-          throw new Error(
-            "The minimum contribution requirement must be 0 or higher"
-          );
-        }
-        if (
-          exchangeVat !== undefined &&
-          (exchangeVat < 0 || exchangeVat > 100 * 100)
-        ) {
-          throw new Error("VAT must be a percentage from 0 to 100");
-        }
-
-        let updated = await prisma.bucket.update({
-          where: { id: bucketId },
-          data: {
-            title,
-            description,
-            summary,
-            ...(typeof budgetItems !== "undefined" && {
-              BudgetItems: {
-                deleteMany: {},
-                createMany: { data: budgetItems },
-              },
-            }),
-            ...(typeof images !== "undefined" && {
-              Images: { deleteMany: {}, createMany: { data: images } },
-            }),
-            directFundingEnabled,
-            directFundingType,
-            exchangeDescription,
-            exchangeMinimumContribution,
-            exchangeVat,
-          },
-          include: {
-            Images: true,
-            FieldValues: true,
-            BudgetItems: true,
-            round: {
-              include: {
-                fields: true,
-                group: {
-                  include: {
-                    discourse: true,
-                    groupMembers: { where: { userId: user.id } },
-                  },
-                },
-              },
-            },
-          },
-        });
-
-        await eventHub.publish("edit-bucket", {
-          currentGroup: updated.round.group,
-          currentGroupMember: updated.round.group?.groupMembers?.[0],
-          round: updated.round,
-          bucket: updated,
-        });
-
-        updated = await updateFundedPercentage(updated);
-
-        return updated;
-      }
-    ),
-    createTag: combineResolvers(
-      isCollModOrAdmin,
-      async (parent, { roundId, tagValue }) => {
-        return await prisma.round.update({
-          where: { id: roundId },
-          data: {
-            tags: {
-              create: {
-                value: tagValue,
-              },
-            },
-          },
-        });
-      }
-    ),
-    addTag: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (parent, { bucketId, tagId }) => {
-        if (!tagId) throw new Error("You need to provide tag id");
-
-        return await prisma.bucket.update({
-          where: { id: bucketId },
-          data: {
-            tags: {
-              connect: {
-                id: tagId,
-              },
-            },
-          },
-        });
-      }
-    ),
-    // removes a tag from all buckets it's added to, and then deletes it
-    deleteTag: combineResolvers(
-      isCollModOrAdmin,
-      async (_, { roundId, tagId }) => {
-        // verify that the tag is part of this round
-        const tag = await prisma.tag.findUnique({
-          where: {
-            id: tagId,
-          },
-        });
-        if (tag?.roundId !== roundId) throw new Error("Incorrect round");
-
-        await prisma.tag.delete({
-          where: { id: tagId },
-        });
-
-        return await prisma.round.findUnique({
-          where: { id: roundId },
-          include: { tags: true },
-        });
-      }
-    ),
-    // removes a tag from a specific bucket
-    removeTag: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId, tagId }) =>
-        prisma.bucket.update({
-          where: { id: bucketId },
-          data: { tags: { disconnect: { id: tagId } } },
-        })
-    ),
     editBucketCustomField: combineResolvers(
       isBucketCocreatorOrCollAdminOrMod,
       async (
@@ -804,1053 +228,11 @@ const resolvers = {
         return updated;
       }
     ),
-    deleteBucket: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId }, { user, eventHub }) => {
-        const {
-          _sum: { amount: contributionsForBucket },
-        } = await prisma.contribution.aggregate({
-          where: { bucketId },
-          _sum: { amount: true },
-        });
-
-        if (contributionsForBucket > 0) {
-          throw new Error(
-            "You cant delete a bucket that has received contributions"
-          );
-        }
-
-        const bucket = await prisma.bucket.update({
-          where: { id: bucketId },
-          data: { deleted: true },
-          include: {
-            round: {
-              include: {
-                group: {
-                  include: { groupMembers: { where: { userId: user.id } } },
-                },
-              },
-            },
-          },
-        });
-
-        await eventHub.publish("delete-bucket", {
-          currentGroup: bucket.round.group,
-          currentGroupMember: bucket.round.group?.groupMembers?.[0],
-          round: bucket.round,
-          bucket: bucket,
-        });
-
-        return bucket;
-      }
-    ),
-    addCocreator: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (parent, { bucketId, memberId }) =>
-        prisma.bucket.update({
-          where: { id: bucketId },
-          data: {
-            cocreators: {
-              connect: { id: memberId },
-            },
-          },
-        })
-    ),
-    removeCocreator: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId, memberId }) =>
-        prisma.bucket.update({
-          where: { id: bucketId },
-          data: { cocreators: { disconnect: { id: memberId } } },
-        })
-    ),
-    publishBucket: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId, unpublish }, { user, eventHub }) => {
-        const bucket = await prisma.bucket.findUnique({
-          where: { id: bucketId },
-          include: {
-            round: {
-              include: {
-                group: {
-                  include: {
-                    groupMembers: { where: { userId: user.id } },
-                    discourse: true,
-                  },
-                },
-              },
-            },
-          },
-        });
-
-        const publishedAt = unpublish ? null : new Date();
-        const resultBucket = await prisma.bucket.update({
-          where: { id: bucket.id },
-          data: { publishedAt },
-        });
-
-        await eventHub.publish("publish-bucket", {
-          currentGroup: bucket.round.group,
-          currentGroupMember: bucket.round.group?.groupMembers?.[0],
-          round: bucket.round,
-          bucket: bucket,
-          unpublish,
-        });
-
-        return resultBucket;
-      }
-    ),
-    addComment: combineResolvers(
-      isCollMember,
-      async (_, { content, bucketId }, { user, eventHub }) => {
-        const bucket = await prisma.bucket.findUnique({
-          where: { id: bucketId },
-          include: {
-            round: {
-              include: {
-                roundMember: { where: { userId: user.id } },
-                group: {
-                  include: {
-                    discourse: true,
-                    groupMembers: { where: { userId: user.id } },
-                  },
-                },
-              },
-            },
-          },
-        });
-        const currentGroup = bucket.round.group;
-        const currentGroupMember = currentGroup?.groupMembers?.[0];
-        const currentCollMember = bucket.round.roundMember?.[0];
-
-        if (
-          groupHasDiscourse(currentGroup) &&
-          !currentGroupMember.discourseApiKey
-        ) {
-          throw new Error(
-            "You need to have a discourse account connected, go to /connect-discourse"
-          );
-        }
-
-        if (content.length < (currentGroup?.discourse?.minPostLength || 3)) {
-          throw new Error(
-            `Your post needs to be at least ${
-              currentGroup.discourse?.minPostLength || 3
-            } characters long!`
-          );
-        }
-
-        const comment = { content, collMemberId: currentCollMember.id };
-
-        const { discourse, prisma: prismaResult } = await eventHub.publish(
-          "create-comment",
-          {
-            currentGroup,
-            currentGroupMember,
-            currentCollMember,
-            currentUser: user,
-            bucket: bucket,
-            round: bucket.round,
-            comment,
-          }
-        );
-
-        return discourse || prismaResult;
-      }
-    ),
-
-    deleteComment: combineResolvers(
-      isCollMember,
-      async (_, { bucketId, commentId }, { user, eventHub }) => {
-        const bucket = await prisma.bucket.findUnique({
-          where: { id: bucketId },
-          include: {
-            comments: { where: { id: commentId } },
-            round: {
-              include: {
-                roundMember: { where: { userId: user.id } },
-                group: {
-                  include: {
-                    groupMembers: { where: { userId: user.id } },
-                    discourse: true,
-                  },
-                },
-              },
-            },
-          },
-        });
-        const currentGroup = bucket.round.group;
-        const currentGroupMember = currentGroup?.groupMembers?.[0];
-        const currentCollMember = bucket.round.roundMember?.[0];
-        const comment = bucket.comments?.[0];
-
-        await eventHub.publish("delete-comment", {
-          currentGroup,
-          currentGroupMember,
-          round: bucket.round,
-          currentCollMember,
-          bucket: bucket,
-          comment,
-        });
-
-        return comment;
-      }
-    ),
-    editComment: combineResolvers(
-      isCollMember,
-      async (parent, { bucketId, commentId, content }, { user, eventHub }) => {
-        let comment = await prisma.comment.findUnique({
-          where: { id: commentId },
-          include: { bucket: { include: { round: true } } },
-        });
-        comment = { ...comment, content };
-
-        const currentCollMember = await prisma.roundMember.findUnique({
-          where: {
-            userId_roundId: {
-              userId: user.id,
-              roundId: comment.bucket.round.id,
-            },
-          },
-          include: {
-            user: true,
-            round: {
-              include: {
-                group: {
-                  include: {
-                    groupMembers: { where: { userId: user.id } },
-                    discourse: true,
-                  },
-                },
-              },
-            },
-          },
-        });
-
-        // TODO: permissions?
-        //if (!roundMember || comment.groupMemberId !== currentGroupMember)
-        const { discourse, prisma: prismaResult } = await eventHub.publish(
-          "edit-comment",
-          {
-            currentGroup: currentCollMember.round.group,
-            currentGroupMember:
-              currentCollMember.round.group?.groupMembers?.[0],
-            currentCollMember,
-            bucket: comment.bucket,
-            comment,
-          }
-        );
-        return discourse || prismaResult;
-      }
-    ),
-    raiseFlag: async (parent, { bucketId, guidelineId, comment }, { user }) => {
-      const currentCollMember = await isAndGetCollMember({
-        bucketId,
-        userId: user.id,
-      });
-
-      // todo: check not already left a flag?
-      const bucket = await prisma.bucket.findUnique({
-        where: { id: bucketId },
-        include: {
-          round: true,
-        },
-      });
-
-      if (!bucket.round.bucketReviewIsOpen || !bucket.publishedAt)
-        throw new Error(
-          "You can only review buckets when bucket review is open and the bucket is published"
-        );
-
-      if (!currentCollMember || !currentCollMember.isApproved)
-        throw new Error("You need to be logged in and/or approved");
-
-      let updated = await prisma.bucket.update({
-        where: { id: bucketId },
-        data: {
-          flags: {
-            create: {
-              guidelineId,
-              type: "RAISE_FLAG",
-              collMemberId: currentCollMember.id,
-              comment,
-            },
-          },
-        },
-        include: {
-          round: {
-            include: {
-              guidelines: { where: { id: guidelineId } },
-              group: { include: { discourse: true } },
-            },
-          },
-        },
-      });
-
-      const logContent = `Someone flagged this bucket for the **${updated.round.guidelines[0].title}** guideline: \n> ${comment}`;
-      const currentGroup = updated.round.group;
-      if (groupHasDiscourse(currentGroup)) {
-        if (!updated.discourseTopicId) {
-          // TODO: break out create thread into separate function
-          const discoursePost = await discourse(
-            currentGroup.discourse
-          ).posts.create(
-            {
-              title: bucket.title,
-              raw: `https://${process.env.DEPLOY_URL}/${currentGroup.slug}/${bucket.round.slug}/${bucket.id}`,
-              ...(currentGroup.discourse.dreamsCategoryId && {
-                category: currentGroup.discourse.dreamsCategoryId,
-              }),
-            },
-            {
-              username: "system",
-            }
-          );
-          updated = await prisma.bucket.update({
-            where: { id: bucketId },
-            data: { discourseTopicId: discoursePost.topic_id },
-            include: {
-              round: {
-                include: {
-                  guidelines: { where: { id: guidelineId } },
-                  group: { include: { discourse: true } },
-                },
-              },
-            },
-          });
-        }
-
-        await discourse(currentGroup.discourse).posts.create(
-          {
-            topic_id: updated.discourseTopicId,
-            raw: logContent,
-          },
-          { username: "system" }
-        );
-      } else {
-        await prisma.comment.create({
-          data: {
-            content: logContent,
-            isLog: true,
-            collMemberId: currentCollMember.id,
-            bucketId,
-          },
-        });
-      }
-
-      return updated;
-    },
-    resolveFlag: async (parent, { bucketId, flagId, comment }, { user }) => {
-      const currentCollMember = await isAndGetCollMember({
-        bucketId,
-        userId: user.id,
-      });
-
-      // todo: check not already left a flag?
-      const bucket = await prisma.bucket.findUnique({
-        where: { id: bucketId },
-        include: {
-          round: true,
-          flags: {
-            where: { id: flagId },
-            include: { guideline: true },
-          },
-        },
-      });
-
-      if (!bucket.round.bucketReviewIsOpen || !bucket.publishedAt)
-        throw new Error(
-          "You can only review buckets when bucket review is open and the bucket is published"
-        );
-
-      if (!currentCollMember || !currentCollMember.isApproved)
-        throw new Error("You need to be logged in and/or approved");
-
-      let updated = await prisma.bucket.update({
-        where: { id: bucketId },
-        data: {
-          flags: {
-            create: {
-              resolvingFlagId: flagId,
-              type: "RESOLVE_FLAG",
-              collMemberId: currentCollMember.id,
-              comment,
-            },
-          },
-        },
-        include: {
-          round: {
-            include: { group: { include: { discourse: true } } },
-          },
-        },
-      });
-      const currentGroup = updated.round.group;
-      const resolvedFlagGuideline = bucket.flags[0].guideline;
-
-      const logContent = `Someone resolved a flag for the **${resolvedFlagGuideline.title}** guideline: \n> ${comment}`;
-
-      if (groupHasDiscourse(currentGroup)) {
-        if (!bucket.discourseTopicId) {
-          // TODO: break out create thread into separate function
-          const discoursePost = await discourse(
-            currentGroup.discourse
-          ).posts.create(
-            {
-              title: bucket.title,
-              raw: `https://${process.env.DEPLOY_URL}/${currentGroup.slug}/${bucket.round.slug}/${bucket.id}`,
-              ...(currentGroup.discourse.dreamsCategoryId && {
-                category: currentGroup.discourse.dreamsCategoryId,
-              }),
-            },
-            {
-              username: "system",
-            }
-          );
-          updated = await prisma.bucket.update({
-            where: { id: bucketId },
-            data: { discourseTopicId: discoursePost.topic_id },
-            include: {
-              round: {
-                include: { group: { include: { discourse: true } } },
-              },
-            },
-          });
-        }
-        await discourse(currentGroup.discourse).posts.create(
-          {
-            topic_id: updated.discourseTopicId,
-            raw: logContent,
-          },
-          { username: "system" }
-        );
-      } else {
-        await prisma.comment.create({
-          data: {
-            content: logContent,
-            isLog: true,
-            collMemberId: currentCollMember.id,
-            bucketId: bucket.id,
-          },
-        });
-      }
-
-      return updated;
-    },
-    allGoodFlag: async (parent, { bucketId }, { user }) => {
-      const currentCollMember = await isAndGetCollMember({
-        bucketId,
-        userId: user.id,
-      });
-
-      const bucket = await prisma.bucket.findUnique({
-        where: { id: bucketId },
-        include: {
-          round: true,
-          flags: {
-            where: {
-              collMemberId: currentCollMember.id,
-              type: "ALL_GOOD_FLAG",
-            },
-          },
-        },
-      });
-
-      if (!bucket.round.bucketReviewIsOpen || !bucket.publishedAt)
-        throw new Error(
-          "You can only review buckets when bucket review is open and the bucket is published"
-        );
-
-      if (bucket.flags.length) {
-        return bucket;
-        // TODO: update the ui to stop the user from doing this. in what way?
-        //throw new Error("You have already left an all good flag");
-      }
-
-      return await prisma.bucket.update({
-        where: { id: bucketId },
-        data: {
-          flags: {
-            create: {
-              type: "ALL_GOOD_FLAG",
-              collMemberId: currentCollMember.id,
-            },
-          },
-        },
-      });
-    },
-
-    joinGroup: async (_, { groupId }, { user }) => {
-      if (!user) throw new Error("You need to be logged in.");
-
-      const group = await prisma.group.findUnique({
-        where: { id: groupId },
-      });
-
-      if (group.registrationPolicy === "INVITE_ONLY")
-        throw new Error("This group is invite only");
-
-      return await prisma.groupMember.create({
-        data: {
-          userId: user.id,
-          groupId: groupId,
-          isApproved: group.registrationPolicy === "OPEN",
-        },
-      });
-    },
-    updateProfile: async (_, { name, username }, { user }) => {
-      if (!user) throw new Error("You need to be logged in..");
-
-      if (!validateUsername(username)) throw new Error("Username is not valid");
-
-      // check case insensitive uniquness of username
-      const existingUser = await prisma.user.findFirst({
-        where: { username: { mode: "insensitive", equals: username } },
-      });
-
-      if (existingUser) throw new Error("Username is already taken");
-
-      return prisma.user.update({
-        where: { id: user.id },
-        data: {
-          name,
-          username,
-        },
-      });
-    },
-    updateBio: async (_, { roundId, bio }, { user }) => {
-      if (!user) throw new Error("You need to be logged in..");
-
-      return prisma.roundMember.update({
-        where: { userId_roundId: { userId: user.id, roundId } },
-        data: {
-          bio,
-        },
-      });
-    },
-    inviteRoundMembers: combineResolvers(
-      isCollOrGroupAdmin,
-      async (_, { emails: emailsString, roundId }, { user: currentUser }) => {
-        const round = await prisma.round.findUnique({
-          where: { id: roundId },
-          include: { group: true },
-        });
-        const emails = emailsString.split(",");
-
-        if (emails.length > 1000)
-          throw new Error("You can only invite 1000 people at a time");
-
-        const invitedRoundMembers = [];
-
-        for (let email of emails) {
-          email = email.trim().toLowerCase();
-
-          const user = await prisma.user.findUnique({
-            where: { email },
-          });
-
-          const updated = await prisma.user.upsert({
-            where: { email },
-            create: {
-              email,
-              collMemberships: {
-                create: {
-                  isApproved: true,
-                  round: { connect: { id: roundId } },
-                  hasJoined: false,
-                  statusAccount: { create: {} },
-                  incomingAccount: { create: {} },
-                  outgoingAccount: { create: {} },
-                },
-              },
-            },
-            update: {
-              collMemberships: {
-                upsert: {
-                  create: {
-                    isApproved: true,
-                    round: { connect: { id: roundId } },
-                    hasJoined: false,
-                    statusAccount: { create: {} },
-                    incomingAccount: { create: {} },
-                    outgoingAccount: { create: {} },
-                  },
-                  update: {
-                    isApproved: true,
-                    isRemoved: false,
-                  },
-                  where: {
-                    userId_roundId: {
-                      userId: user?.id ?? "undefined",
-                      roundId,
-                    },
-                  },
-                },
-              },
-            },
-            include: { collMemberships: { where: { roundId } } },
-          });
-
-          await emailService.inviteMember({ email, currentUser, round });
-
-          invitedRoundMembers.push(updated.collMemberships?.[0]);
-        }
-        return invitedRoundMembers;
-      }
-    ),
-    inviteGroupMembers: combineResolvers(
-      isGroupAdmin,
-      async (_, { groupId, emails: emailsString }, { user: currentUser }) => {
-        const emails: string[] = emailsString.split(",");
-
-        if (emails.length > 1000)
-          throw new Error("You can only invite 1000 people at a time");
-
-        const newGroupMembers = [];
-
-        for (let email of emails) {
-          email = email.trim().toLowerCase();
-
-          const user = await prisma.user.upsert({
-            where: {
-              email,
-            },
-            create: {
-              groupMemberships: { create: { groupId: groupId } },
-              email,
-            },
-            update: {
-              groupMemberships: {
-                create: {
-                  groupId: groupId,
-                },
-              },
-            },
-            include: {
-              groupMemberships: {
-                where: { groupId: groupId },
-                include: { group: true },
-              },
-            },
-          });
-          const groupMembership = user.groupMemberships?.[0];
-          const currentGroup = groupMembership.group;
-
-          await emailService.inviteMember({ email, currentUser, currentGroup });
-
-          newGroupMembers.push(groupMembership);
-        }
-
-        return newGroupMembers;
-      }
-    ),
-    updateGroupMember: combineResolvers(
-      isGroupAdmin,
-      async (parent, { groupId, memberId, isAdmin, isApproved }, { user }) => {
-        const groupMember = await prisma.groupMember.findFirst({
-          where: { id: memberId, groupId: groupId },
-        });
-
-        if (!groupMember) throw new Error("No member to update found");
-
-        if (typeof isAdmin !== "undefined") {
-          if (isAdmin === false) {
-            const groupAdmins = await prisma.groupMember.findMany({
-              where: { groupId: groupId, isAdmin: true },
-            });
-            if (groupAdmins.length <= 1)
-              throw new Error("You need at least 1 group admin");
-          }
-          if (typeof isAdmin !== "undefined") groupMember.isAdmin = isAdmin;
-        }
-
-        if (typeof isApproved !== "undefined")
-          groupMember.isApproved = isApproved;
-
-        return await prisma.groupMember.update({
-          where: { id: groupMember.id },
-          data: { ...groupMember },
-        });
-      }
-    ),
-    deleteGroupMember: combineResolvers(
-      isGroupAdmin,
-      async (parent, { groupId, groupMemberId }) => {
-        return prisma.groupMember.delete({
-          where: { id: groupMemberId },
-        });
-      }
-    ),
-    updateMember: combineResolvers(
-      isCollOrGroupAdmin,
-      async (
-        parent,
-        { roundId, memberId, isApproved, isAdmin, isModerator }
-      ) => {
-        const roundMember = await prisma.roundMember.findFirst({
-          where: { roundId, id: memberId },
-        });
-        if (!roundMember)
-          throw new Error("This member does not exist in this round");
-
-        return prisma.roundMember.update({
-          where: { id: memberId },
-          data: {
-            isApproved,
-            isAdmin,
-            isModerator,
-          },
-        });
-      }
-    ),
-    deleteMember: combineResolvers(
-      isCollOrGroupAdmin,
-      async (parent, { roundId, memberId }) => {
-        const roundMember = await prisma.roundMember.findUnique({
-          where: { id: memberId },
-        });
-        if (!roundMember)
-          throw new Error("This member does not exist in this collection");
-
-        if ((await roundMemberBalance(roundMember)) !== 0) {
-          throw new Error(
-            "You can only remove a round participant with 0 balance"
-          );
-        }
-
-        return prisma.roundMember.update({
-          where: { id: memberId },
-          data: {
-            isApproved: false,
-            hasJoined: false,
-            isRemoved: true,
-            isAdmin: false,
-            isModerator: false,
-          },
-        });
-      }
-    ),
-    // deleteGroup: async (parent, { groupId }, { user }) => {
-    //   const { currentGroupMember } = await getCurrentGroupAndMember({
-    //     groupId: groupId,
-    //     user,
-    //   });
-
-    //   if (
-    //     !(
-    //       (currentGroupMember &&
-    //         currentGroupMember.isAdmin &&
-    //         groupId == currentGroupMember.groupId) ||
-    //       user.isRootAdmin
-    //     )
-    //   )
-    //     throw new Error(
-    //       "You need to be group. or root admin to delete an group"
-    //     );
-    //   //TODO: turn into soft delete
-    //   return prisma.group.delete({ where: { id: groupId } });
-    // },
-    approveForGranting: combineResolvers(
-      async (parent, args, ctx) => {
-        const round = await prisma.round.findFirst({
-          where: { buckets: { some: { id: args.bucketId } } },
-        });
-
-        return round.requireBucketApproval
-          ? isCollModOrAdmin(parent, args, ctx)
-          : isBucketCocreatorOrCollAdminOrMod(parent, args, ctx);
-      },
-      async (_, { bucketId, approved }) =>
-        prisma.bucket.update({
-          where: { id: bucketId },
-          data: {
-            approvedAt: approved ? new Date() : null,
-            ...(approved && { canceledAt: null }),
-          },
-        })
-    ),
-    allocate: async (_, { roundMemberId, amount, type }, { user }) => {
-      const targetRoundMember = await prisma.roundMember.findUnique({
-        where: { id: roundMemberId },
-      });
-
-      const currentCollMember = await prisma.roundMember.findUnique({
-        where: {
-          userId_roundId: {
-            userId: user.id,
-            roundId: targetRoundMember.roundId,
-          },
-        },
-      });
-
-      if (!currentCollMember?.isAdmin)
-        throw new Error("You are not admin for this round");
-
-      await prisma.$transaction(async (prisma) => {
-        await allocateToMember({
-          member: targetRoundMember,
-          roundId: targetRoundMember.roundId,
-          amount,
-          type,
-          allocatedBy: currentCollMember.id,
-          prisma: prisma as any,
-        });
-      });
-
-      return targetRoundMember;
-    },
-    bulkAllocate: combineResolvers(
-      isCollOrGroupAdmin,
-      async (_, { roundId, amount, type }, { user }) => {
-        const roundMembers = await prisma.roundMember.findMany({
-          where: {
-            roundId: roundId,
-            isApproved: true,
-          },
-        });
-        //here
-        const currentCollMember = await prisma.roundMember.findUnique({
-          where: {
-            userId_roundId: {
-              userId: user.id,
-              roundId: roundId,
-            },
-          },
-        });
-
-        await bulkAllocate({
-          roundId,
-          amount,
-          type,
-          allocatedBy: currentCollMember.id,
-        });
-
-        return roundMembers;
-      }
-    ),
     contribute: async (_, { roundId, bucketId, amount }, { user }) => {
       return contribute({ roundId, bucketId, amount, user });
     },
-    markAsCompleted: combineResolvers(
-      isCollModOrAdmin,
-      async (_, { bucketId }) =>
-        prisma.bucket.update({
-          where: { id: bucketId },
-          data: { completedAt: new Date() },
-        })
-    ),
-    acceptFunding: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId }) => {
-        const {
-          _sum: { amount: contributionsForBucket },
-        } = await prisma.contribution.aggregate({
-          where: { bucketId },
-          _sum: { amount: true },
-        });
-
-        const {
-          _sum: { min: minExpenses },
-        } = await prisma.budgetItem.aggregate({
-          where: { bucketId, type: "EXPENSE" },
-          _sum: { min: true },
-        });
-
-        const {
-          _sum: { min: minIncome },
-        } = await prisma.budgetItem.aggregate({
-          where: { bucketId, type: "INCOME" },
-          _sum: { min: true },
-        });
-
-        const minGoal = minIncome - minExpenses;
-
-        if (contributionsForBucket < minGoal)
-          throw new Error("Bucket has not reached its minimum goal yet.");
-
-        return prisma.bucket.update({
-          where: { id: bucketId },
-          data: { fundedAt: new Date() },
-        });
-      }
-    ),
-    cancelFunding: combineResolvers(
-      isBucketCocreatorOrCollAdminOrMod,
-      async (_, { bucketId }, { eventHub }) => {
-        const bucket = await prisma.bucket.findUnique({
-          where: { id: bucketId },
-          include: {
-            cocreators: true,
-            round: { include: { group: true } },
-            Contributions: {
-              include: {
-                roundMember: {
-                  include: { user: { include: { emailSettings: true } } },
-                },
-              },
-            },
-          },
-        });
-
-        if (bucket.completedAt)
-          throw new Error(
-            "This bucket has already been marked completed, can't cancel funding."
-          );
-
-        const updated = await prisma.bucket.update({
-          where: { id: bucketId },
-          data: {
-            fundedAt: null,
-            approvedAt: null,
-            canceledAt: new Date(),
-            Contributions: { deleteMany: {} },
-            statusAccount: {
-              update: { incomingTransactions: { deleteMany: {} } },
-            },
-          },
-        });
-
-        await eventHub.publish("cancel-funding", {
-          bucket,
-        });
-
-        return updated;
-      }
-    ),
-    updateGrantingSettings: combineResolvers(
-      isCollOrGroupAdmin,
-      async (
-        parent,
-        {
-          roundId,
-          currency,
-          maxAmountToBucketPerUser,
-          bucketCreationCloses,
-          grantingOpens,
-          grantingCloses,
-          allowStretchGoals,
-          requireBucketApproval,
-          directFundingEnabled,
-          directFundingTerms,
-        }
-      ) => {
-        const round = await prisma.round.findUnique({
-          where: { id: roundId },
-        });
-        const grantingHasOpened = dayjs(round.grantingOpens).isBefore(dayjs());
-
-        if (currency && grantingHasOpened) {
-          throw new Error(
-            "You can't change currency after funding has started"
-          );
-        }
-
-        if (directFundingEnabled && !(await stripeIsConnected({ round }))) {
-          throw new Error("You need to connect this round to Stripe first");
-        }
-
-        return prisma.round.update({
-          where: { id: roundId },
-          data: {
-            currency,
-            maxAmountToBucketPerUser,
-            bucketCreationCloses,
-            grantingOpens,
-            grantingCloses,
-            allowStretchGoals,
-            requireBucketApproval,
-            directFundingEnabled,
-            directFundingTerms,
-          },
-        });
-      }
-    ),
-    acceptInvitation: async (parent, { roundId }, { user }) => {
-      if (!user) throw new Error("You need to be logged in.");
-
-      const member = await getRoundMember({
-        roundId,
-        userId: user.id,
-      });
-
-      if (!member) {
-        throw new Error("You are not a participant in this round");
-      }
-
-      if (member.hasJoined) {
-        throw new Error("Invitation not pending");
-      }
-
-      return prisma.roundMember.update({
-        where: { id: member.id },
-        data: {
-          hasJoined: true,
-        },
-      });
-    },
-    joinRound: async (parent, { roundId }, { user }) => {
-      if (!user) throw new Error("You need to be logged in.");
-
-      const currentGroupMember = await prisma.groupMember.findFirst({
-        where: {
-          userId: user.id,
-          group: { rounds: { some: { id: roundId } } },
-        },
-      });
-
-      const round = await prisma.round.findUnique({
-        where: { id: roundId },
-      });
-
-      if (
-        !currentGroupMember?.isAdmin &&
-        round.registrationPolicy === "INVITE_ONLY"
-      )
-        throw new Error("This round is invite only");
-
-      const isApproved =
-        currentGroupMember?.isAdmin || round.registrationPolicy === "OPEN";
-
-      const roundMember = await prisma.roundMember.upsert({
-        where: { userId_roundId: { userId: user.id, roundId } },
-        create: {
-          round: { connect: { id: roundId } },
-          user: { connect: { id: user.id } },
-          isApproved,
-          statusAccount: { create: {} },
-          incomingAccount: { create: {} },
-          outgoingAccount: { create: {} },
-        },
-        update: { isApproved, hasJoined: true, isRemoved: false },
-      });
-
-      if (!isApproved) {
-        await emailService.roundJoinRequest({
-          round,
-          roundMember,
-        });
-      }
-
-      return roundMember;
-    },
-    setEmailSetting: async (parent, { settingKey, value }, { user }) => {
-      if (!user) throw "You need to be logged in";
-
-      await prisma.emailSettings.upsert({
-        where: { userId: user.id },
-        create: { userId: user.id, [settingKey]: value },
-        update: { [settingKey]: value },
-      });
-
-      return prisma.user.findUnique({ where: { id: user.id } });
-    },
-    acceptTerms: async (parent, args, { user }) => {
-      if (!user) throw new Error("You need to be logged in");
-      return prisma.user.update({
-        where: { id: user.id },
-        data: { acceptedTermsAt: new Date() },
-      });
-    },
   },
+
   RoundMember: {
     round: async (member) => {
       return await prisma.round.findUnique({
