+++ ui/server/graphql/resolvers/mutations/index.ts
@@ -1,0 +1,4 @@
+export * as userMutations from "./user";
+export * as groupMutations from "./group";
+export * as roundMutations from "./round";
+export * as bucketMutations from "./bucket";
