+++ ui/server/graphql/resolvers/queries/index.ts
@@ -1,0 +1,4 @@
+export * as userQueries from "./user";
+export * as groupQueries from "./group";
+export * as roundQueries from "./round";
+export * as bucketQueries from "./bucket";
