--- ui/server/graphql/resolvers/round/index.ts
@@ -1,1 +1,0 @@
-export * as queries from "./queries";
